@@ -3,7 +3,10 @@ package cn.codepub.algorithms.graph;
 import cn.codepub.algorithms.graph.utils.UnionFindSet;
 import org.junit.Test;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.PriorityQueue;
+import java.util.Queue;
 
 /**
  * <p>
@@ -26,16 +29,14 @@ public class Kruskal {
     /**
      * 实现一个匿名类，提供了基于边的权值的比较器
      */
-    public static final Queue<Edge> QUEUE = new PriorityQueue<Edge>(new Comparator<Edge>() {
-        @Override
-        public int compare(Edge e1, Edge e2) {
-            if (e1.value == e2.value) {
-                return 0;
-            } else {
-                return e1.value > e2.value ? 1 : -1;
-            }
+    public static final Queue<Edge> QUEUE = new PriorityQueue<>((e1, e2) -> {
+        if (e1.value == e2.value) {
+            return 0;
+        } else {
+            return e1.value > e2.value ? 1 : -1;
         }
     });
+
     private static List<Edge> visitedEdges = new ArrayList<>();//已访问的边集
     private static List<Edge> edgeList = new ArrayList<>();//边集
 
